EES-2170 EES-2301 Fix broken UI tests and address other comments of PR #2760

+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/Methodologies/MethodologyService.cs
@@ -125,7 +125,8 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.Methodologie
             return await _persistenceHelper.CheckEntityExists<Methodology>(id, queryable =>
                     queryable.Include(m => m.MethodologyParent)
                         .ThenInclude(mp => mp.Publications))
-                .OnSuccess(_userService.CheckCanUpdateMethodology)
+                .OnSuccess(methodology => _userService.CheckCanApproveMethodology(methodology)
+                    .OrElse(() => _userService.CheckCanMarkMethodologyAsDraft(methodology)))
                 .OnSuccess(async methodology =>
                 {
                     // Get all Publications using the Methodology including adopting Publications
@@ -153,52 +154,61 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.Methodologie
         {
             return await _persistenceHelper
                 .CheckEntityExists<Methodology>(id, HydrateMethodologyForManageMethodologySummaryViewModel)
-                .OnSuccess(methodology => methodology.Status != request.Status
-                    ? UpdateMethodologyStatus(methodology, request)
-                    : UpdateMethodologyDetails(methodology, request))
+                .OnSuccess(methodology => UpdateMethodologyStatus(methodology, request))
+                .OnSuccess(methodology => UpdateMethodologyDetails(methodology, request))
                 .OnSuccess(_ => GetSummary(id));
         }
 
-        private Task<Either<ActionResult, Methodology>> UpdateMethodologyStatus(Methodology methodologyToUpdate,
+        private async Task<Either<ActionResult, Methodology>> UpdateMethodologyStatus(Methodology methodologyToUpdate,
             MethodologyUpdateRequest request)
         {
-            return
-                CheckCanUpdateMethodologyStatus(methodologyToUpdate, request.Status)
-                    .OnSuccessDo(methodology => CheckMethodologyCanDependOnRelease(methodology, request))
-                    .OnSuccessDo(RemoveUnusedImages)
-                    .OnSuccess(async methodology =>
+            if (!request.IsStatusUpdateForMethodology(methodologyToUpdate))
+            {
+                // Status unchanged
+                return methodologyToUpdate;
+            }
+
+            return await CheckCanUpdateMethodologyStatus(methodologyToUpdate, request.Status)
+                .OnSuccessDo(methodology => CheckMethodologyCanDependOnRelease(methodology, request))
+                .OnSuccessDo(RemoveUnusedImages)
+                .OnSuccess(async methodology =>
+                {
+                    methodology.Status = request.Status;
+                    methodology.PublishingStrategy = request.PublishingStrategy;
+                    methodology.ScheduledWithReleaseId = request.WithReleaseId;
+                    methodology.InternalReleaseNote = Approved == request.Status
+                        ? request.LatestInternalReleaseNote
+                        : null;
+
+                    methodology.Updated = DateTime.UtcNow;
+
+                    _context.Methodologies.Update(methodology);
+
+                    if (await _methodologyRepository.IsPubliclyAccessible(methodology.Id))
                     {
-                        methodology.Status = request.Status;
-                        methodology.PublishingStrategy = request.PublishingStrategy;
-                        methodology.ScheduledWithReleaseId = request.WithReleaseId;
-                        methodology.InternalReleaseNote = Approved == request.Status
-                            ? request.LatestInternalReleaseNote
-                            : null;
+                        methodology.Published = DateTime.UtcNow;
 
-                        methodology.Updated = DateTime.UtcNow;
+                        await _publishingService.PublishMethodologyFiles(methodology.Id);
 
-                        _context.Methodologies.Update(methodology);
+                        // Invalidate the 'All Methodologies' cache item
+                        await _blobCacheService.DeleteItem(new AllMethodologiesCacheKey());
+                    }
 
-                        if (await _methodologyRepository.IsPubliclyAccessible(methodology.Id))
-                        {
-                            methodology.Published = DateTime.UtcNow;
-
-                            await _publishingService.PublishMethodologyFiles(methodology.Id);
-
-                            // Invalidate the 'All Methodologies' cache item
-                            await _blobCacheService.DeleteItem(new AllMethodologiesCacheKey());
-                        }
-
-                        await _context.SaveChangesAsync();
-                        return methodology;
-                    });
+                    await _context.SaveChangesAsync();
+                    return methodology;
+                });
         }
 
-        private Task<Either<ActionResult, Methodology>> UpdateMethodologyDetails(Methodology methodologyToUpdate,
+        private async Task<Either<ActionResult, Methodology>> UpdateMethodologyDetails(Methodology methodologyToUpdate,
             MethodologyUpdateRequest request)
         {
-            return _userService
-                .CheckCanUpdateMethodology(methodologyToUpdate)
+            if (!request.IsDetailUpdateForMethodology(methodologyToUpdate))
+            {
+                // Details unchanged
+                return methodologyToUpdate;
+            }
+
+            return await _userService.CheckCanUpdateMethodology(methodologyToUpdate)
                 // Check that the Methodology will have a unique slug.  It is possible to have a clash in the case where
                 // another Methodology has previously set its AlternativeTitle (and Slug) to something specific and then
                 // this Methodology attempts to set its AlternativeTitle (and Slug) to the same value.  Whilst an
@@ -324,19 +334,13 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.Methodologie
         }
 
         private Task<Either<ActionResult, Methodology>> CheckCanUpdateMethodologyStatus(Methodology methodology,
-            MethodologyStatus status)
+            MethodologyStatus requestedStatus)
         {
-            if (methodology.Status == status)
-            {
-                // Status unchanged
-                return Task.FromResult(new Either<ActionResult, Methodology>(methodology));
-            }
-
-            return status switch
+            return requestedStatus switch
             {
                 Draft => _userService.CheckCanMarkMethodologyAsDraft(methodology),
                 Approved => _userService.CheckCanApproveMethodology(methodology),
-                _ => throw new ArgumentOutOfRangeException(nameof(status), "Unexpected status")
+                _ => throw new ArgumentOutOfRangeException(nameof(requestedStatus), "Unexpected status")
             };
         }
